+++ integration_tests/elemental_11.py
@@ -4,7 +4,7 @@ from numpy import empty, arctanh, reshape, float64, sinh, sqrt, sin,  cosh
 def verify1d_arctanh(array: f32[:], result: f32[:], size: i32):
     i: i32
     eps: f32
-    eps = 1e-6
+    eps = f32(1e-6)
 
     for i in range(size):
         assert abs(arctanh(array[i]) - result[i]) <= eps
@@ -19,7 +19,7 @@ def verifynd_arctanh(array: f64[:, :, :], result: f64[:, :, :], size1: i32, size
     for i in range(size1):
         for j in range(size2):
             for k in range(size3):
-                assert abs( arctanh(array[i, j, k])**2 - result[i, j, k]) <= eps
+                assert abs( arctanh(array[i, j, k])**2.0 - result[i, j, k]) <= eps
 
 
 def elemental_arctanh():
@@ -31,7 +31,7 @@ def elemental_arctanh():
     arctanh1d: f32[999] = empty(999)
 
     for i in range(999):
-        array1d[i] = ((-1)**i) * (float(i)/1000.0)
+        array1d[i] = f32(f64((-1)**i) * (float(i)/1000.0))
 
     arctanh1d = arctanh(array1d)
     verify1d_arctanh(array1d, arctanh1d, 999)
@@ -42,9 +42,9 @@ def elemental_arctanh():
     for i in range(100):
         for j in range(50):
             for k in range(10):
-                arraynd[i, j, k] = sin(((-1)**k) * float(i + j + k))
+                arraynd[i, j, k] = sin(f64((-1)**k) * float(i + j + k))
 
-    arctanhnd =  arctanh(arraynd)**2
+    arctanhnd =  arctanh(arraynd)**2.0
     verifynd_arctanh(arraynd, arctanhnd, 100, 50, 10)
 
 
@@ -67,8 +67,8 @@ def elemental_trig_identity():
                 for l in range(4):
                     arraynd[i, j, k, l] = ((float(i + j + k + l)))/25.0
 
-    identity1 = 2 * arctanh(arraynd) - arctanh((2 * arraynd) / ( 1 + arraynd**2))
-    identity2 =  cosh(arctanh(arraynd)) - (sqrt(1 - (arraynd**2)))**(-1)
+    identity1 = 2.0 * arctanh(arraynd) - arctanh((2.0 * arraynd) / ( 1.0 + arraynd**2.0))
+    identity2 =  cosh(arctanh(arraynd)) - (sqrt(1.0 - (arraynd**2.0)))**(-1.0)
 
     newshape: i32[1] = empty(1, dtype=int)
     newshape[0] = 400
