+++ integration_tests/structs_01.py
@@ -1,4 +1,4 @@
-from ltypes import i32, f32, dataclass
+from ltypes import i32, f32, f64, dataclass
 
 @dataclass
 class A:
@@ -11,22 +11,22 @@ def f(a: A):
 
 def change_struct(a: A):
     a.x = a.x + 1
-    a.y = a.y + 1
+    a.y = a.y + f32(1)
 
 def g():
     x: A
-    x = A(3.25, 3)
+    x = A(f32(3.25), 3)
     f(x)
     assert x.x == 3
-    assert x.y == 3.25
+    assert f64(x.y) == 3.25
 
     x.x = 5
-    x.y = 5.5
+    x.y = f32(5.5)
     f(x)
     assert x.x == 5
-    assert x.y == 5.5
+    assert f64(x.y) == 5.5
     change_struct(x)
     assert x.x == 6
-    assert x.y == 6.5
+    assert f64(x.y) == 6.5
 
 g()
