+++ integration_tests/structs_02.py
@@ -11,13 +11,13 @@ def f(a: CPtr) -> None:
     y: f32
     a1: A
     a2: Pointer[A]
-    a1 = A(3, 3.25)
+    a1 = A(3, f32(3.25))
     a2 = pointer(a1)
     print(a2, pointer(a1))
     x = a2.x
     y = a2.y
     assert x == 3
-    assert y == 3.25
+    assert f64(y) == 3.25
     c_p_pointer(a, a2)
     print(a, a2, pointer(a1))
